Split paragraphs in individual lines
Markdown treats `\n` not as a new line when rendered to HTML. Instead, the text continues in the same paragraph. (For a `<br>` you have to add a empty line (`\n\n`))

This made reading the changelog very hard in the rendered mode. I added lists where appropriate.

And yes, Github handles this differently for comments in issues/PRs